+++ src/compiler/Libraries/PackageGenerator/Base/ArcSymbolBase.cs
@@ -4,7 +4,7 @@ namespace Arc.Compiler.PackageGenerator.Base
 {
     public abstract class ArcSymbolBase
     {
-        public long Id { get; internal set; } = new Random().NextInt64();
+        public ulong Id { get; internal set; } = (ulong)new Random().NextInt64();
 
         public required string Name { get; set; }
